Added stage virtual functions
